Safari-611.1.10/LayoutTests/fast/custom-elements/exceptions-for-synchronous-custom-element-creation-expected.txt

CONSOLE MESSAGE: TypeError: The result of constructing a custom element must be a HTMLElement
CONSOLE MESSAGE: NotSupportedError: A newly constructed custom element must not have attributes
CONSOLE MESSAGE: NotSupportedError: A newly constructed custom element must not have child nodes
CONSOLE MESSAGE: NotSupportedError: A newly constructed custom element must not have a parent node
CONSOLE MESSAGE: NotSupportedError: A newly constructed custom element belongs to a wrong document
CONSOLE MESSAGE: NotSupportedError: A newly constructed custom element has incorrect local name
